Update batterynotif(uname).sh
updated check charger status before shutdown

+++ batterynotif(uname).sh
@@ -23,9 +23,21 @@ do
     	then
        		/usr/bin/notify-send -i "$ICONC" "About to shutdown connect charger" "Battery level is ${battery_level}%!" 
 	 	paplay /usr/share/sounds/freedesktop/stereo/dialog-warning.oga
+		
          	sleep 30s
+		STATUS=$(cat $POWERSUPPLY)
+		if [ $STATUS = 1 ]
+		then
+			continue
+		fi
+		
 		/usr/bin/notify-send -I "$ICONS" "Too late, Shutting down"
 		sleep 15s
+		STATUS=$(cat $POWERSUPPLY)
+		if [ $STATUS = 1 ]
+		then
+			continue
+		fi
 	 	systemctl poweroff 
 
 	elif [ $battery_level -le 15 -a $STATUS = $NOT_CHARGING ] #battery low 2 final warning
